Interactive-Deep-Reasearch/requirements.txt
# Interactive Deep Research - 依赖清单

# 核心依赖
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
pydantic>=2.0.0

# 配置管理
pyyaml>=6.0

# 工具依赖
requests>=2.31.0
numpy>=1.24.0

# 开发依赖
pytest>=7.4.0
pytest-asyncio>=0.21.0
jupyter>=1.0.0

# 可选依赖
# streamlit>=1.28.0  # 用于Web界面
# matplotlib>=3.7.0  # 用于数据可视化